requirements.txt
# 邮件总结工具核心依赖
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
python-dotenv>=0.19.0
pydantic>=2.0.0

# 邮件处理相关
imapclient>=2.3.1
email-validator>=1.3.0
beautifulsoup4>=4.11.0
html2text>=2020.1.16

# 数据处理
xmltodict>=0.13.0
requests>=2.28.0
typing-extensions>=4.4.0

# 图像处理（用于邮件附件）
Pillow>=9.0.0